feat(app): cicd tab
