+++ src/ui/shared/index.ts
@@ -65,3 +65,4 @@ export * from "./key-value";
 export * from "./org-requirements";
 export * from "./survey";
 export * from "./env-editor";
+export * from "./app-auto-deploy-guide";
